requirements.txt
Flask
Werkzeug
click==8.0
Flask-SQLAlchemy==3.0.3
Flask-Login
psycopg2
MarkupSafe
WTForms==3.0.1
FLASK_WTF==1.1.1
email-validator==1.3.1